Refactor assistance measures customization
