+++ frontend/src/lib-customizations/espoo/featureFlags.tsx
@@ -3,7 +3,6 @@
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
 export default {
-  assistanceMeasureChildServiceEnabled: true,
   urgencyAttachmentsEnabled: true,
   preschoolEnabled: true
 }
